pyproject.toml
[project]
name = "trading-service"
version = "0.1.0"
description = "Lighter DEX pair trading service with admin panel"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "pydantic-settings>=2.0.0",
    "apscheduler>=3.10.0",
    "cryptography>=42.0.0",
    "hyperliquid-python-sdk>=0.3.0",
    "lighter-sdk",
    "pandas>=2.0.0",
    "numpy>=1.24.0,<2.0.0",
    "statsmodels>=0.14.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pyotp>=2.9.0",
    "qrcode[pil]>=7.4.2",
    "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
